server/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Vector Database
chromadb==0.4.22

# AI APIs
groq==0.4.1
# openai==1.10.0  # Optional: only if using OpenAI embeddings

# Embeddings
sentence-transformers==2.3.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
numpy==1.26.3
